Add Error Prone check: SelfAssignment
This may seem like an overreaction on Error Prone's part, but it turns
out that this was deliberate choice on their part: see
https://github.com/google/error-prone/issues/2196#issuecomment-798802207
where they say that they found that rewriting the (non-erroneous)
self-casts into something more directly solving the problem is clearer
to the reader anyway. So instead of writing immediately to the field,
store the value in a local variable first.