Enable Error Prone check: CheckedExceptionNotThrown
@kokosing keeps wishing that this was on.

The issue with this is that:

* This is an experimental checker
* It has a huge number of false negatives

Regarding the second point, it may be that making it more accurate would
make it unviably slow (the equivalent inspection in IntelliJ IDEA takes
a lot of time, has to look at the whole code base - it's definitely not
local - and has to be done repeatedly until nothing more is found). But
perhaps it's still useful.